requirements.txt
attrs==20.3.0
marshmallow==3.5.1
webargs>=5.5.3
pytest==5.4.1
twine==3.1.1
wheel==0.34.2
sqlalchemy==1.3.17
